(С ОШИБКАМИ) Перенес ClawSh
