+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/Claw.java
@@ -7,46 +7,18 @@ import com.qualcomm.robotcore.hardware.Servo;
 @Config
 public class Claw {
 
-    private final Servo clawServoShoulder;
     private final Servo clawServoLift;
     private final LinearOpMode opMode;
 
-    public static double CLAW_CLOSE = 0.055;
-    public static double CLAW_OPEN = 0.26;
     public static double CLAW_OPEN_LIFT = 0.45;
     public static double CLAW_CLOSE_LIFT = 0.145;
-
-    public static double CLAW_HALF_OPEN = 0.6;
-
-    public boolean stateOpenShoulder;
     public boolean stateOpenLift;
 
 
     public Claw(LinearOpMode opMode) {
         this.opMode = opMode;
-        clawServoShoulder = opMode.hardwareMap.servo.get("ClawServoShoulder");
         clawServoLift = opMode.hardwareMap.servo.get("ClawServoLift");
     }
-
-    public void halfOpenSh() {
-        clawServoShoulder.setPosition(CLAW_HALF_OPEN);
-    }
-
-    public void halfOpenLift() {
-        clawServoLift.setPosition(CLAW_HALF_OPEN);
-    }
-
-    public void switchPositionShoulder() {
-        if (!stateOpenShoulder) {
-            clawServoShoulder.setPosition(CLAW_OPEN);
-            stateOpenShoulder = true;
-        } else {
-            clawServoShoulder.setPosition(CLAW_CLOSE);
-            stateOpenShoulder = false;
-        }
-    }
-
-
 
     public void switchPositionLift() {
         if (!stateOpenLift) {
@@ -58,27 +30,16 @@ public class Claw {
         }
     }
 
-
-
-    public void closeSh() {
-        clawServoShoulder.setPosition(CLAW_CLOSE);
-        stateOpenShoulder = false;
-    }
-    public void openSh() {
-        clawServoShoulder.setPosition(CLAW_OPEN);
-        stateOpenShoulder = true;
-
-    }
     public void closeLift() {
-        clawServoShoulder.setPosition(CLAW_CLOSE_LIFT);
-        stateOpenShoulder = false;
+        clawServoLift.setPosition(CLAW_CLOSE_LIFT);
+        stateOpenLift = false;
     }
     public void openLift() {
         clawServoLift.setPosition(CLAW_OPEN_LIFT);
-        stateOpenShoulder = true;
+        stateOpenLift = true;
     }
     public void strongCloseSh() {
-        clawServoShoulder.setPosition(0.05);
+        clawServoLift.setPosition(0.05);
     }
 }
