+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/test/TeleopClaw.java
@@ -4,6 +4,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.modules.Claw;
+import org.firstinspires.ftc.teamcode.modules.Shoulder;
 
 @TeleOp(name = "Claw", group = "Test")
 public class TeleopClaw extends LinearOpMode {
@@ -15,6 +16,7 @@ public class TeleopClaw extends LinearOpMode {
     @Override
     public void runOpMode() {
         Claw cl = new Claw(this);
+        Shoulder sl = new Shoulder(this);
 
         while (opModeInInit()) {
 
@@ -23,14 +25,11 @@ public class TeleopClaw extends LinearOpMode {
         while (opModeIsActive()){
 
             if (gamepad2.a) {
-                cl.halfOpenSh();
-            }
-            if (gamepad2.x) {
-                cl.halfOpenLift();
+                sl.halfOpenSh();
             }
 
             if (gamepad2.left_bumper && !btnStateSh) {
-                cl.switchPositionShoulder();
+                sl.switchPositionShoulder();
             }
             if (gamepad2.right_bumper && !btnStateLift) {
                 cl.switchPositionLift();
@@ -38,7 +37,7 @@ public class TeleopClaw extends LinearOpMode {
             btnStateSh = gamepad2.left_bumper;
             btnStateLift = gamepad2.right_bumper;
 
-            telemetry.addData("Значение клешни на плече: ", cl.stateOpenShoulder);
+            telemetry.addData("Значение клешни на плече: ", sl.stateOpenShoulder);
             telemetry.addData("Значение клешни на подъемнике: ", cl.stateOpenLift);
             telemetry.addLine("Управление:");
             telemetry.addLine(" А - Полуоткрытое состояние (плечо)");
